backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
networkx==3.2.1
numpy==1.24.3
scikit-learn==1.3.0
scipy==1.11.1
requests==2.31.0
python-dotenv==1.0.0
